requirements.txt
numpy==1.26.0
einops==0.7.0
tqdm==4.66.1
ipython==8.17.2
dm_control==1.0.10
mujoco==2.3.2
gym==0.26.2
scipy==1.11.3
tensorboard==2.15.1
omegaconf==2.3.0
hydra-core==1.3.2
moviepy==1.0.3
termcolor==2.3.0
imageio==2.27 # bug in 2.28 make the gif still
